requirements.txt
aniso8601==9.0.1
cffi==1.14.6
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flask-swagger-ui==3.36.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.3
python-dotenv==0.19.0
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1

